requirements.txt
streamlit==1.28.1
openai==1.3.0
Pillow==10.0.1
pandas==2.1.1
python-dotenv==1.0.0